Ticket #415: B integer types are moved to uint32_t. Remove corresponding useless casts.

+++ csrc/services/b2c/service_mgr_bs.c
@@ -116,7 +116,7 @@ void service_mgr_bs__client_channel_connected_event_discovery(
                     if (elt != NULL)
                     {
                         SOPC_Services_EnqueueEvent(APP_TO_SE_SEND_DISCOVERY_REQUEST,
-                                                   (uint32_t) service_mgr_bs__channel_config_idx, elt->msgToSend,
+                                                   service_mgr_bs__channel_config_idx, elt->msgToSend,
                                                    elt->msgAppContext);
                     }
                     free(elt);
